Ignore SIGINT whilst the pager is running
Some pagers (notably less) bind SIGINT to actions such as closing a
prompt, so we shouldn't exit upon receiving SIGINT whilst the pager is
running. If the pager does exit as a result of a SIGINT, we'll get an
EPIPE the next time we try to send a line to it anyway, and exit as a
result.